+++ app/src/main/java/org/simple/clinic/facilitypicker/FacilityPickerView.kt
@@ -134,7 +134,6 @@ class FacilityPickerView(
 
   override fun onDetachedFromWindow() {
     delegate.stop()
-    binding = null
     super.onDetachedFromWindow()
   }