+++ app/src/main/java/org/simple/clinic/summary/bloodsugar/view/BloodSugarSummaryView.kt
@@ -192,7 +192,6 @@ class BloodSugarSummaryView(
 
   override fun onDetachedFromWindow() {
     delegate.stop()
-    binding = null
     super.onDetachedFromWindow()
   }