+++ app/src/main/java/org/simple/clinic/summary/nextappointment/NextAppointmentCardView.kt
@@ -129,7 +129,6 @@ class NextAppointmentCardView(
 
   override fun onDetachedFromWindow() {
     delegate.stop()
-    binding = null
     super.onDetachedFromWindow()
   }