+++ app/src/main/java/org/simple/clinic/summary/prescribeddrugs/DrugSummaryView.kt
@@ -117,7 +117,6 @@ class DrugSummaryView(
 
   override fun onDetachedFromWindow() {
     delegate.stop()
-    binding = null
     super.onDetachedFromWindow()
   }