+++ app/src/main/java/org/simple/clinic/teleconsultlog/prescription/medicines/TeleconsultMedicinesView.kt
@@ -138,7 +138,6 @@ class TeleconsultMedicinesView(
 
   override fun onDetachedFromWindow() {
     super.onDetachedFromWindow()
-    binding = null
     delegate.stop()
   }