+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Internal
 
+## 2023-08-07-8827
+
+### Internal
+
 - Bump AGP to v8.1.0
 - Stop clearing view binding reference when a custom view is detached
 - Add missing SQLCipher proguard rules for R8